requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pkg-resources==0.0.0
pytz==2019.3
PyYAML==3.13
six==1.13.0
Werkzeug==0.16.0
